% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mch.R
\name{calc_mch}
\alias{calc_mch}
\title{Calculate Mean Corpuscular Hemoglobin (MCH)}
\usage{
calc_mch(hb, rb)
}
\arguments{
\item{hb}{Numeric value representing the hemoglobin concentration in g/dL.}

\item{rb}{Numeric value representing the red blood cell count (millions/ µL).}
}
\value{
Numeric value representing the MCH in picograms (pg).
}
\description{
Calculates the MCH based on hemoglobin concentration and red blood cell count.
MCH measures the average amount of hemoglobin per red blood cell.
}
\examples{
calc_mch(15, 5)
}
