% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mchc.R
\name{calc_mchc}
\alias{calc_mchc}
\title{Calculate Mean Corpuscular Hemoglobin Concentration (MCHC)}
\usage{
calc_mchc(hb, ht)
}
\arguments{
\item{hb}{Numeric value representing the hemoglobin concentration in g/dL.}

\item{ht}{Numeric value representing the hematocrit in percentage.}
}
\value{
Numeric value representing the MCHC in grams per deciliter (g/dL).
}
\description{
Calculates the MCHC based on hemoglobin concentration and hematocrit.
MCHC measures the average concentration of hemoglobin in red blood cells.
}
\examples{
calc_mchc(15, 40)
}
