% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mcv.R
\name{calc_mcv}
\alias{calc_mcv}
\title{Calculate Mean Corpuscular Volume (MCV)}
\usage{
calc_mcv(ht, rb)
}
\arguments{
\item{ht}{Numeric value representing the hematocrit in percentage.}

\item{rb}{Numeric value representing the red blood cell count (millions/ µL).}
}
\value{
Numeric value representing the MCV in femtoliters (fL).
}
\description{
MCV measures the average volume of red blood cells in a blood sample.
}
\examples{
calc_mcv(40, 5)
}
